+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Added preliminary support for Python 3.13 (#1686)
 
+- `py_is_null_xptr()` and `[[` now load delayed modules (#1688).
+
 - Fixed error when attempting to use a python venv created with `uv` (#1678)
 
 - Fixed error where `py_discover_config()` attempted to detect